Changed package names, added marker interfaces instead of hardcoding package names, changed return type of methods in PersonController, used better way in EntityManagerFactory for defining package names.
