+++ .gitignore
@@ -1,4 +1,7 @@
 *.class
+*.classpath
+*.project
+/.settings
 
 # Mobile Tools for Java (J2ME)
 .mtj.tmp/
@@ -10,3 +13,4 @@
 
 # virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
 hs_err_pid*
+/target/
