+++ src/main/java/eu/execom/labs/test_integration_showcase/configuration/RootConfig.java
@@ -1,11 +1,13 @@
-package eu.execom.representation.spring_mvc_integration.configuration;
+package eu.execom.labs.test_integration_showcase.configuration;
 
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
 
+import eu.execom.labs.test_integration_showcase.service.ServiceComponents;
+
 @Configuration
 @Import({DataConfig.class})
-@ComponentScan(basePackages = "eu.execom.representation.spring_mvc_integration.service")
+@ComponentScan(basePackageClasses = ServiceComponents.class)
 public class RootConfig {
 }
